Introducing SvelteFlow.
Can do vertices and edges.
Can do double binding and in-situ mapping.

Mapping from matrix to graph works already.

Mapping from graph to matrix has to be implemented (basically mapping
Routes to the matrix representation; The algorithms are straightforward
and I've done them for the C/C++ client, too, but a bit tedious in JS
and no time left right now).

That is, this is half way in progress :)