Pass globals via index.html.
This makes subsequent changing easier, cf. #8.

Tried to fix the Gitlab CI pipeline in case of distributed workers, but did not do this "fiddling in index.html with sed and friends" because it also does not look the right way of doing it. This way we are somehow half the way of doing #8.